+++ handlers/facebook/facebook.go
@@ -231,6 +231,8 @@ func (h *handler) receiveEvents(ctx context.Context, channel courier.Channel, w 
 	// the list of data we will return in our response
 	data := make([]interface{}, 0, 2)
 
+	seenMsgIDs := make(map[string]bool, 2)
+
 	// for each entry
 	for _, entry := range payload.Entry {
 		// no entry, ignore
@@ -350,6 +352,9 @@ func (h *handler) receiveEvents(ctx context.Context, channel courier.Channel, w 
 
 		} else if msg.Message != nil {
 			// this is an incoming message
+			if seenMsgIDs[msg.Message.MID] {
+				continue
+			}
 
 			// ignore echos
 			if msg.Message.IsEcho {
@@ -385,8 +390,7 @@ func (h *handler) receiveEvents(ctx context.Context, channel courier.Channel, w 
 			}
 
 			// create our message
-			ev := h.Backend().NewIncomingMsg(channel, urn, text, clog).WithExternalID(msg.Message.MID).WithReceivedOn(date)
-			event := h.Backend().CheckExternalIDSeen(ev)
+			event := h.Backend().NewIncomingMsg(channel, urn, text, msg.Message.MID, clog).WithReceivedOn(date)
 
 			// add any attachment URL found
 			for _, attURL := range attachmentURLs {
@@ -398,10 +402,9 @@ func (h *handler) receiveEvents(ctx context.Context, channel courier.Channel, w 
 				return nil, err
 			}
 
-			h.Backend().WriteExternalIDSeen(event)
-
 			events = append(events, event)
 			data = append(data, courier.NewMsgReceiveData(event))
+			seenMsgIDs[msg.Message.MID] = true
 
 		} else if msg.Delivery != nil {
 			// this is a delivery report