+++ handlers/macrokiosk/macrokiosk.go
@@ -127,10 +127,8 @@ func (h *handler) receiveMessage(ctx context.Context, channel courier.Channel, w
 		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, err)
 	}
 
-	// build our msg
-	msg := h.Backend().NewIncomingMsg(channel, urn, form.Text, clog).WithExternalID(form.MsgID).WithReceivedOn(date.UTC())
-
-	// and write it
+	// create and write the message
+	msg := h.Backend().NewIncomingMsg(channel, urn, form.Text, form.MsgID, clog).WithReceivedOn(date.UTC())
 	return handlers.WriteMsgsAndResponse(ctx, h, []courier.Msg{msg}, w, r, clog)
 }