+++ handlers/mblox/mblox.go
@@ -96,7 +96,7 @@ func (h *handler) receiveEvent(ctx context.Context, channel courier.Channel, w h
 		}
 
 		// build our Message
-		msg := h.Backend().NewIncomingMsg(channel, urn, payload.Body, clog).WithReceivedOn(date.UTC()).WithExternalID(payload.ID)
+		msg := h.Backend().NewIncomingMsg(channel, urn, payload.Body, payload.ID, clog).WithReceivedOn(date.UTC())
 
 		// and finally write our message
 		return handlers.WriteMsgsAndResponse(ctx, h, []courier.Msg{msg}, w, r, clog)